+++ Changelog.md
@@ -1,4 +1,7 @@
 
+## 2025-05-07
+ * allow puppets without a builder
+
 ## 2025-05-04
  * skip hands 'None' only for the current scene
  * allow '. ' in prop names